+++ src/Services/EmailService.js
@@ -1,15 +1,16 @@
 const dotenv = require("dotenv")
 dotenv.config()
 
+
 const jwt = require("jsonwebtoken")
 const mailer = require("nodemailer")
 
+const { GenerateEmailToken } = require("./../Modules/Tokens")
+
 async function SendVerifyEmail(email, id){
-    let token = jwt.sign({
-        id: id
-    }, process.env.JWT_EMAIL_SECRET, {algorithm: "HS512", expiresIn: "15m"})
-    //send token to given email
-    //dunno why but even the test account doesnt want to work 
+    let token = await GenerateEmailToken(email, id)
+    console.log(token)
+    return true
 }
 
 module.exports = {