+++ src/config.js
@@ -3,8 +3,14 @@ dotenv.config();
 
 module.exports = {
     DATABASE_URL: process.env.DATABASE_URL,
-    JWT_SECRET: process.env.JWT_SECRET,
-    JWT_REFRESH_SECRET: process.env.JWT_REFRESH_SECRET,
-    JWT_EMAIL_SECRET: process.env.JWT_EMAIL_SECRET,
+
+    ACCESS_TOKEN_SECRET: process.env.ACCESS_TOKEN_SECRET,
+    ACCESS_TOKEN_EXPIRE: process.env.ACCESS_TOKEN_EXPIRE,
+    REFRESH_TOKEN_SECRET: process.env.REFRESH_TOKEN_SECRET,
+    REFRESH_TOKEN_EXPIRE: process.env.REFRESH_TOKEN_EXPIRE,
+    MAX_REFRESH_TOKENS: process.env.MAX_REFRESH_TOKENS,
+    EMAIL_TOKEN_SECRET: process.env.EMAIL_TOKEN_SECRET,
+    EMAIL_TOKEN_EXPIRE: process.env.EMAIL_TOKEN_EXPIRE,
+
     PORT: Number(process.env.PORT) || 8080
 }